requirements.txt
paho-mqtt
streamlit